backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.110.3
greenlet==3.0.3
h11==0.14.0
idna==3.7
jose==1.0.0
passlib==1.7.4
pyasn1==0.6.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
PyMySQL==1.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlmodel==0.0.18
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
numpy==1.23.3
opencv-python
ultralytics
Pillow==9.3.0      
cryptography